Refactor ExtractAudioMetadata: move helpers to src/core/lib
Split the monolithic ExtractAudioMetadata.mjs (1,550 lines) into four
focused files totalling 990 lines:

- src/core/lib/AudioBytes.mjs: byte-reading and text-decoding utilities
- src/core/lib/AudioMetaSchema.mjs: report skeleton and container detection
- src/core/lib/AudioParsers.mjs: all 9 format parsers and internal helpers
- src/core/operations/ExtractAudioMetadata.mjs: operation class only

Also replaces infoURL with a Wikipedia link per review feedback.

No functional changes — all 1,805 tests pass, ESLint clean.
